Refactor map loading and point cloud extraction to navigation submodule
